Fix 'contains' pattern for substring 'includes'
