+++ dropwizard-metrics/src/main/java/io/dropwizard/metrics/DefaultStringMatchingStrategy.java
@@ -1,0 +1,16 @@
+package io.dropwizard.metrics;
+
+import com.google.common.collect.ImmutableSet;
+
+class DefaultStringMatchingStrategy implements StringMatchingStrategy {
+    @Override
+    public boolean containsMatch(ImmutableSet<String> matchExpressions, String metricName) {
+        for (String matchExpression : matchExpressions) {
+            if (metricName.contains(matchExpression)) {
+                // just need to match on a single value - return as soon as we do
+                return true;
+            }
+        }
+        return false;
+    }
+}
