+++ dropwizard-metrics/src/main/java/io/dropwizard/metrics/StringMatchingStrategy.java
@@ -1,0 +1,7 @@
+package io.dropwizard.metrics;
+
+import com.google.common.collect.ImmutableSet;
+
+interface StringMatchingStrategy {
+    boolean containsMatch(ImmutableSet<String> matchExpressions, String metricName);
+}
